[OpenXR] Implement DeviceDelegateOpenXR::UpdateClockLevels()
Wolvic can ask the system to boost CPU/GPU performance for faster
rendering when entering a VR experience or when watching an immersive
video. Also we could enter in a more battery saving mode when not
requiring a lot of performance (like when in 2D browsing).
That was implemented in the OculusVR backend but never done
in the OpenXR one, this is a regression from the migration.

In order to do that we use the XR_EXT_performance_settings extension
whenever available. That OpenXR extension allows us to set the
performance settings of both CPU and GPU.