[OpenXR] Fix reorientation for 3DoF devices
We recently added support to reorientation changes in the OpenXR
backend. It was working perfectly for 6DoF devices however in
the case of 3DoF devices the widget position was locked to head's.
That's because we were re-using the firstPose field to check
whether a recentering happened (as we knew it's set to null on
those events). However it's also set to null in other cases
(EnterVR for example), and also not properly updated in
3DoF devices. So we better use a different field to store
that information.