Pass a color argument to renderPointerColor() to change it dynamically
In the presence of layers, the Pointer class is using an Android surface
to draw the circles, and that happens on Java's side. In that method,
the circle colors are hard-coded to Black (outer) and White (inner).

We want to be able to change the inner circle's color dynamically to
other colors for certain conditions (e.g, a pinch gesture during hand
tracking); so we need to pass a color argument to that method.

Apart from this new functionality, the current feature of changing
the pointer color in Settings is broken. It only changes the beams
color but not the pointer target. This patch fixes that.